re-arranged wallet interface functions
