+++ app/walletcore/wallet.go
@@ -86,11 +86,6 @@ type Wallet interface {
 
 	// ChangePrivatePassphrase changes the private passphrase from the oldPass to the provided newPass
 	ChangePrivatePassphrase(ctx context.Context, oldPass, newPass string) error
-	// BestBlock fetches the best block on the network
-	BestBlock() (uint32, error)
-
-	// GetConnectedPeersCount returns the number of connected peers
-	GetConnectedPeersCount() int32
 
 	// NetType returns the network type of this wallet
 	NetType() string
