+++ app/walletmediums/dcrlibwallet/walletfunctions.go
@@ -367,10 +367,6 @@ func (lib *DcrWalletLib) ChangePrivatePassphrase(_ context.Context, oldPass, new
 	return lib.walletLib.ChangePrivatePassphrase([]byte(oldPass), []byte(newPass))
 }
 
-func (lib *DcrWalletLib) BestBlock() (uint32, error) {
-	return uint32(lib.walletLib.GetBestBlock()), nil
-}
-
 func (lib *DcrWalletLib) NetType() string {
 	return lib.activeNet.Params.Name
 }
